refactor: notification 구조 변경 (제안) - member 엔티티 구조를 기준으로 notification 엔티티 구조 정리 - 통일을 위해 최소한의 폴더 구조만 유지 - 불필요한 mapper / adapter 계층 제거 (필요 시 확장)
