+++ src/entities/notification/index.ts
@@ -6,13 +6,10 @@ import { dto } from '@/entities/notification/infrastructure';
  * ────────────────────────────────────────────────────*/
 export * from './infrastructure';
 export * from './core';
-export * from './mapper';
 
 /* ─────────────────────────────────────────────────────
  * 스키마
  * ────────────────────────────────────────────────────*/
-export * from './schema';
-
 export const notification = {
   dto,
   domain,
