+++ src/entities/notification/infrastructure/index.ts
@@ -1,4 +1,3 @@
-export { dto } from './notification.dto.schema';
-export { adapters } from './notification.adapters';
+export { dto } from './notification.dto';
 export { repository } from './notification.repository';
 export { notificationKeys } from './notification.keys';
