--- src/entities/notification/infrastructure/notification.adapters.ts
@@ -1,15 +1,0 @@
-import { sharedSchema } from '@/types';
-
-import { dto } from './notification.dto.schema';
-
-/* ─────────────────────────────────────────────────────
- * API 응답 데이터를 도메인 객체로 변환 및 검증
- * ────────────────────────────────────────────────────*/
-const toDomainFromAPI = sharedSchema.response(dto.schema.array());
-
-/* ─────────────────────────────────────────────────────
- * 내보내기
- * ────────────────────────────────────────────────────*/
-export const adapters = {
-  fromApi: toDomainFromAPI,
-};
