+++ src/entities/notification/infrastructure/notification.repository.ts
@@ -1,12 +1,13 @@
-import { notificationMapper } from '@/entities/notification/mapper';
+import { domain } from '@/entities/notification/core';
 import {
   FrontendNotification,
   NotificationDTO,
 } from '@/entities/notification/types';
 import { api } from '@/shared/api';
+import { unwrapEnvelope } from '@/shared/lib/api-utils';
 import { CommonResponse } from '@/types';
 
-import { adapters } from './notification.adapters';
+import { dto } from './notification.dto';
 
 /* ─────────────────────────────────────────────────────
  * [Read] 알림 목록 조회
@@ -16,8 +17,8 @@ const getNotificationList = async (): Promise<FrontendNotification[]> => {
     await api.private.get<CommonResponse<NotificationDTO[]>>(
       '/notification/all'
     );
-  const validatedResponse = adapters.fromApi.parse(response);
-  return notificationMapper.toDomainList(validatedResponse.data ?? []);
+  const dtos = unwrapEnvelope(response, dto.arraySchema);
+  return dtos.map((dto) => domain.schema.parse(dto));
 };
 
 /* ─────────────────────────────────────────────────────
@@ -25,8 +26,8 @@ const getNotificationList = async (): Promise<FrontendNotification[]> => {
  * ────────────────────────────────────────────────────*/
 const getUnreadNotifications = async () => {
   const response = await api.private.get('/notification/unread');
-  const validatedResponse = adapters.fromApi.parse(response);
-  return notificationMapper.toDomainList(validatedResponse.data ?? []);
+  const dtos = unwrapEnvelope(response, dto.arraySchema);
+  return dtos.map((dto) => domain.schema.parse(dto));
 };
 
 /* ─────────────────────────────────────────────────────
