--- src/entities/notification/mapper/index.ts
@@ -1,1 +1,0 @@
-export { notificationMapper } from './notification.mapper';
